ENH: reorganize yaml_ as plugins
