+++ hutch_python/plugins/load.py
@@ -1,4 +1,4 @@
-from . import utils
+from .. import utils
 
 
 def load_objs(info):
