+++ hutch_python/plugins/namespace.py
@@ -1,7 +1,7 @@
 from types import SimpleNamespace
 import logging
 
-from . import utils
+from .. import utils
 
 logger = logging.getLogger(__name__)
